Split out viewmodels and removed quantity from Ingredient view
