+++ srcs/Food/Models/Shared/IngredientQuantityViewModel.cs
@@ -1,0 +1,15 @@
+﻿using Common.EnumDataTypes;
+
+namespace Food.Models.Ingredients
+{
+    public class IngredientQuantityViewModel
+    {
+        public decimal Quantity { get; set; }
+        public QuantityType QuantityType { get; set; }
+
+        public override string ToString()
+        {
+            return $"{Quantity}{QuantityType}";
+        }
+    }
+}
